+++ Datamigratie.Server/Features/Map/ZaaktypeMapping/MapZaaktypeDetails/RoltypeMapping/ShowRoltypeMappings/Models/RoltypeMappingResponse.cs
@@ -1,0 +1,7 @@
+﻿namespace Datamigratie.Server.Features.Map.ZaaktypeMapping.MapZaaktypeDetails.RoltypeMapping.ShowRoltypeMappings.Models;
+
+public class RoltypeMappingResponse
+{
+    public required string DetRol { get; set; }
+    public required string OzRoltypeUrl { get; set; }
+}