+++ Datamigratie.Server/Features/Map/ZaaktypeMapping/MapZaaktypeDetails/ShowOzZaaktypeInfo/Models/EnrichedOzZaaktype.cs
@@ -1,4 +1,4 @@
-using Datamigratie.Common.Services.OpenZaak.Models;
+﻿using Datamigratie.Common.Services.OpenZaak.Models;
 using Datamigratie.Server.Shared.Models;
 
 namespace Datamigratie.Server.Features.Map.ZaaktypeMapping.MapZaaktypeDetails.ShowOzZaaktypeInfo.Models
@@ -12,6 +12,8 @@ namespace Datamigratie.Server.Features.Map.ZaaktypeMapping.MapZaaktypeDetails.Sh
         public List<OzInformatieobjecttype> Informatieobjecttypen { get; set; } = [];
 
         public List<OzBesluittype> Besluittypen { get; set; } = [];
+
+        public List<OzRoltype> Roltypen { get; set; } = [];
 
         public List<ZaaktypeOptionItem> OzZaakVertrouwelijkheidaanduidingen { get; set; } = [];