+++ Datamigratie.Tests/Server/Features/MigrateZaken/ManageMigrations/StartMigration/StartSingleMigration/StartSingleMigrationControllerTests.cs
@@ -51,7 +51,8 @@ public class StartSingleMigrationControllerTests
         DocumenttypeMappings = [],
         ZaakVertrouwelijkheidMappings = [],
         BesluittypeMappings = [],
-        PdfInformatieobjecttypeId = Guid.Empty
+        PdfInformatieobjecttypeId = Guid.Empty,
+        RoltypeMappings = []
     };
 
     private static StartSingleMigrationController CreateSut(