+++ docs/changelog/changelog.md
@@ -9,6 +9,8 @@
 - [Enable performance testing using large data volumes](https://dimpact.atlassian.net/browse/DATA-253)
 - [Allow partial re-run of migration](https://dimpact.atlassian.net/browse/DATA-297)
 - [(QoL) Auto fill all zaaktype mappings with UI button](https://dimpact.atlassian.net/browse/DATA-407)
+- [Enable single zaak migration from Zaaktype page](https://dimpact.atlassian.net/browse/DATA-265)
+- [Configuration of Roltype per Zaaktype](https://dimpact.atlassian.net/browse/DATA-331)
 - [Extend PDF content: un-mapped properties](https://dimpact.atlassian.net/browse/DATA-386)
 
 ## v0.1.0